lodcat.gensim/requirements.txt
gensim == 4.1.2
javaobj-py3 == 0.4.3
